Split entity allocator into a dedicated module
