+++ src/lib.rs
@@ -62,15 +62,17 @@ macro_rules! smaller_tuples_too {
 mod archetype;
 mod borrow;
 mod bundle;
+mod entities;
 mod entity_builder;
 mod query;
 mod world;
 
 pub use borrow::{EntityRef, Ref, RefMut};
 pub use bundle::{Bundle, DynamicBundle, MissingComponent};
+pub use entities::{Entity, NoSuchEntity};
 pub use entity_builder::{BuiltEntity, EntityBuilder};
 pub use query::{Query, QueryBorrow, QueryIter, With, Without};
-pub use world::{Component, ComponentError, Entity, Iter, NoSuchEntity, World};
+pub use world::{Component, ComponentError, Iter, World};
 
 // Unstable implementation details needed by the macros
 #[doc(hidden)]
